handle berkleydb for ubuntu20

+++ installNG.sh
@@ -131,33 +131,41 @@ fi
 
 function bdb() {
 
-if [ ! -f "/usr/lib/libdb-4.8.so" ]; then
 
-	dpkg -r libdb-dev
-	cd
-	rm db-4.8.30.NC.tar.gz
-	wget http://download.oracle.com/berkeley-db/db-4.8.30.NC.tar.gz
-	echo '12edc0df75bf9abd7f82f821795bcee50f42cb2e5f76a6a281b85732798364ef  db-4.8.30.NC.tar.gz' | sha256sum -c
-	tar -xvf db-4.8.30.NC.tar.gz
-	cd db-4.8.30.NC/build_unix
-	mkdir -p build
-	BDB_PREFIX=$(pwd)/build
-	../dist/configure --disable-shared --enable-cxx --with-pic --prefix=$BDB_PREFIX
-	make install
-	cd ../..
-	
-fi
+if [[ $(lsb_release -rs) == "20.04" ]]; then 
+	chmod +x /root/handleberkely.sh
+	/root/handleberkely.sh
 
-ln -s /usr/local/BerkeleyDB.4.8 /usr/include/db4.8
-ln -s /usr/include/db4.8/include/* /usr/include
-ln -s /usr/include/db4.8/lib/* /usr/lib
+else
 
-export BDB_INCLUDE_PATH="/usr/local/BerkeleyDB.4.8/include"
-export BDB_LIB_PATH="/usr/local/BerkeleyDB.4.8/lib"
-ln -s /usr/local/BerkeleyDB.4.8/lib/libdb-4.8.so /usr/lib/libdb-4.8.so
+	if [ ! -f "/usr/lib/libdb-4.8.so" ]; then
 
-PATH=$PATH:/usr/include/db4.8
-export PATH
+		dpkg -r libdb-dev
+		cd
+		rm db-4.8.30.NC.tar.gz
+		wget http://download.oracle.com/berkeley-db/db-4.8.30.NC.tar.gz
+		echo '12edc0df75bf9abd7f82f821795bcee50f42cb2e5f76a6a281b85732798364ef  db-4.8.30.NC.tar.gz' | sha256sum -c
+		tar -xvf db-4.8.30.NC.tar.gz
+		cd db-4.8.30.NC/build_unix
+		mkdir -p build
+		BDB_PREFIX=$(pwd)/build
+		../dist/configure --disable-shared --enable-cxx --with-pic --prefix=$BDB_PREFIX
+		make install
+		cd ../..
+
+	fi
+
+	ln -s /usr/local/BerkeleyDB.4.8 /usr/include/db4.8
+	ln -s /usr/include/db4.8/include/* /usr/include
+	ln -s /usr/include/db4.8/lib/* /usr/lib
+
+	export BDB_INCLUDE_PATH="/usr/local/BerkeleyDB.4.8/include"
+	export BDB_LIB_PATH="/usr/local/BerkeleyDB.4.8/lib"
+	ln -s /usr/local/BerkeleyDB.4.8/lib/libdb-4.8.so /usr/lib/libdb-4.8.so
+
+	PATH=$PATH:/usr/include/db4.8
+	export PATH
+fi	
 
 }
